+++ pages/boards/index.tsx
@@ -15,8 +15,6 @@ BoardsPageWithStore.getInitialProps = async (ctx) => {
   const reduxStore = setOrGetStore();
   const { dispatch } = reduxStore;
 
-  await dispatch(fetchBoards());
-
   const userDetails = isValidUser(ctx);
 
   if (userDetails && !userDetails.isValid && typeof window === 'undefined') {
@@ -32,6 +30,8 @@ BoardsPageWithStore.getInitialProps = async (ctx) => {
   if (ctx.req) {
     await dispatch(updateUserData({ type: 'id', value: userDetails && userDetails.id }));
   }
+
+  await dispatch(fetchBoards());
 
   return {
     initialReduxStore: reduxStore.getState()
