+++ src/slices/boards.ts
@@ -1,6 +1,7 @@
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
 import checkEnvironment from '@/util/check-environment';
 import { SingleUser } from '@/src/types/user';
+import { BoardSlice } from '@/src/types/boards';
 
 const initialState = {
   boards: [],
@@ -18,7 +19,39 @@ export const fetchBoards = createAsyncThunk('boards/fetchBoards', async (_obj, {
   const response = await fetch(`${host}/api/boards?userid=${id}`).then((response) =>
     response.json()
   );
+
   return response;
+});
+
+export const createBoard = createAsyncThunk('board/create', async (_obj, { getState }) => {
+  const { board } = getState() as { board: BoardSlice };
+  const { user } = getState() as { user: SingleUser };
+
+  const data = {
+    _id: board.board._id,
+    name: board.board.name,
+    dateCreated: board.board.dateCreated,
+    createdBy: user.id,
+    columns: []
+  };
+
+  const url = `${host}/api/boards`;
+
+  const response = await fetch(url, {
+    method: 'POST',
+    mode: 'cors',
+    cache: 'no-cache',
+    credentials: 'same-origin',
+    headers: {
+      'Content-Type': 'application/json'
+    },
+    redirect: 'follow',
+    referrerPolicy: 'no-referrer',
+    body: JSON.stringify(data)
+  });
+
+  const inJSON = await response.json();
+  return inJSON;
 });
 
 export const boardSlice = createSlice({
@@ -36,6 +69,15 @@ export const boardSlice = createSlice({
       state.status = 'success';
     },
     [fetchBoards.rejected.toString()]: (state) => {
+      state.status = 'failed';
+    },
+    [createBoard.pending.toString()]: (state) => {
+      state.status = 'pending';
+    },
+    [createBoard.fulfilled.toString()]: (state) => {
+      state.status = 'success';
+    },
+    [createBoard.rejected.toString()]: (state) => {
       state.status = 'failed';
     }
   }
